# models/openapi/gateway.openapi.json
{
  "components": {
    "schemas": {
      "CallbackData": {
        "additionalProperties": false,
        "description": "Completed transcription payload delivered to a client callback URL.",
        "properties": {
          "text": {
            "description": "Final transcription text.",
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "CallbackData",
        "type": "object"
      },
      "CallbackError": {
        "additionalProperties": false,
        "description": "Stable error payload delivered when transcription fails.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/CallbackErrorCode",
            "description": "Machine-readable failure code."
          },
          "message": {
            "description": "Safe human-readable failure description.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "CallbackError",
        "type": "object"
      },
      "CallbackErrorCode": {
        "description": "Terminal transcription failures safe to disclose to callback receivers.",
        "enum": [
          "audio_rejected",
          "audio_retrieval_failed",
          "invalid_audio",
          "model_unavailable",
          "transcription_failed",
          "internal_error"
        ],
        "title": "CallbackErrorCode",
        "type": "string"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "The error response body for public gateway operations.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/GatewayError"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "GatewayError": {
        "additionalProperties": false,
        "description": "Stable machine- and human-readable HTTP error detail.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/GatewayErrorCode",
            "description": "Machine-readable failure code."
          },
          "message": {
            "description": "Safe human-readable failure description.",
            "title": "Message",
            "type": "string"
          },
          "param": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Request field responsible for the error, when applicable.",
            "title": "Param"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Gateway request identifier, when available.",
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "GatewayError",
        "type": "object"
      },
      "GatewayErrorCode": {
        "description": "Errors returned directly by the gateway HTTP boundary.",
        "enum": [
          "authentication_required",
          "forbidden",
          "idempotency_conflict",
          "rate_limited",
          "internal_error",
          "model_unavailable"
        ],
        "title": "GatewayErrorCode",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "additionalProperties": false,
        "description": "Liveness response for the gateway process.",
        "properties": {
          "status": {
            "const": "ok",
            "description": "The gateway process is healthy.",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "ModelDescriptor": {
        "additionalProperties": false,
        "description": "Public identity of a model served by this gateway.",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/TranscriptionModel",
            "description": "Internal model identifier accepted by transcription requests."
          }
        },
        "required": [
          "id"
        ],
        "title": "ModelDescriptor",
        "type": "object"
      },
      "ModelList": {
        "additionalProperties": false,
        "description": "Response listing public model identifiers.",
        "properties": {
          "data": {
            "description": "Models currently accepted by this gateway.",
            "items": {
              "$ref": "#/components/schemas/ModelDescriptor"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "ModelList",
        "type": "object"
      },
      "QueuedTranscription": {
        "additionalProperties": false,
        "description": "Immediate response for a transcription accepted for asynchronous work.",
        "properties": {
          "id": {
            "description": "Stable identifier for the accepted transcription.",
            "examples": [
              "transcription_01J0EXAMPLE"
            ],
            "title": "Id",
            "type": "string"
          },
          "report_id": {
            "description": "Opaque caller report identifier supplied in the submission.",
            "examples": [
              "report_01J0EXAMPLE"
            ],
            "title": "Report Id",
            "type": "string"
          },
          "status": {
            "const": "queued",
            "description": "The gateway accepted the request and queued it for processing.",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "report_id",
          "status"
        ],
        "title": "QueuedTranscription",
        "type": "object"
      },
      "TranscriptionCompleted": {
        "additionalProperties": false,
        "description": "At-least-once event sent after successful transcription.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CallbackData"
          },
          "id": {
            "description": "Transcription ID from the queued response.",
            "title": "Id",
            "type": "string"
          },
          "model": {
            "$ref": "#/components/schemas/TranscriptionModel",
            "description": "Model that processed the audio."
          },
          "report_id": {
            "description": "Opaque caller report identifier supplied in the submission.",
            "title": "Report Id",
            "type": "string"
          },
          "type": {
            "const": "transcription.completed",
            "description": "Event type discriminator.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "id",
          "report_id",
          "model",
          "data"
        ],
        "title": "TranscriptionCompleted",
        "type": "object"
      },
      "TranscriptionFailed": {
        "additionalProperties": false,
        "description": "At-least-once event sent when transcription cannot complete.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/CallbackError"
          },
          "id": {
            "description": "Transcription ID from the queued response.",
            "title": "Id",
            "type": "string"
          },
          "model": {
            "$ref": "#/components/schemas/TranscriptionModel",
            "description": "Model selected by the request."
          },
          "report_id": {
            "description": "Opaque caller report identifier supplied in the submission.",
            "title": "Report Id",
            "type": "string"
          },
          "type": {
            "const": "transcription.failed",
            "description": "Event type discriminator.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "id",
          "report_id",
          "model",
          "error"
        ],
        "title": "TranscriptionFailed",
        "type": "object"
      },
      "TranscriptionModel": {
        "description": "Models available through the gateway's stable public contract.",
        "enum": [
          "granite-4.0-1b-speech"
        ],
        "title": "TranscriptionModel",
        "type": "string"
      },
      "TranscriptionRequest": {
        "additionalProperties": false,
        "description": "Request that submits one already-uploaded audio object for transcription.",
        "properties": {
          "audio_url": {
            "description": "HTTPS presigned URL for the audio object. The retrieval worker validates DNS, network, size, and object policy before download.",
            "examples": [
              "https://uploads.example.com/audio.wav?X-Amz-Signature=example"
            ],
            "maxLength": 8192,
            "title": "Audio Url",
            "type": "string"
          },
          "callback_url": {
            "description": "HTTPS endpoint for at-least-once completion or failure delivery. Receivers must deduplicate events by transcription ID.",
            "examples": [
              "https://client.example.com/webhooks/transcriptions"
            ],
            "maxLength": 8192,
            "title": "Callback Url",
            "type": "string"
          },
          "idempotency_key": {
            "description": "Client-generated key for safe retries. The eventual submission store must scope it to the authenticated caller and return the original job for an identical retry.",
            "examples": [
              "customer-request-123"
            ],
            "maxLength": 255,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "model": {
            "$ref": "#/components/schemas/TranscriptionModel",
            "description": "Internal identifier of the model that will process the audio.",
            "examples": [
              "granite-4.0-1b-speech"
            ]
          },
          "report_id": {
            "description": "Opaque caller report identifier. It is retained with the job and echoed in the accepted response and every callback event for correlation.",
            "examples": [
              "report_01J0EXAMPLE"
            ],
            "maxLength": 255,
            "minLength": 1,
            "title": "Report Id",
            "type": "string"
          }
        },
        "required": [
          "model",
          "audio_url",
          "callback_url",
          "report_id",
          "idempotency_key"
        ],
        "title": "TranscriptionRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Submit presigned audio URLs for asynchronous transcription.",
    "title": "SocialGuard ASR Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/healthz": {
      "get": {
        "description": "Return process liveness without checking downstream dependencies.",
        "operationId": "healthCheck",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Check gateway health"
      }
    },
    "/v1/models": {
      "get": {
        "description": "List public internal model identifiers for the authenticated caller.",
        "operationId": "listModels",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelList"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authenticated caller is not permitted to perform the operation."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An idempotency key was reused with a different request payload."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The gateway cannot accept more work at this time."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected gateway failure; implementation details are omitted."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The selected model or submission dependency is unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List accepted models"
      }
    },
    "/v1/transcriptions": {
      "post": {
        "callbacks": {
          "transcription_callback": {
            "{$request.body#/callback_url}": {
              "post": {
                "description": "The gateway delivers exactly one terminal event type at least once and retries until a configured retry window expires. Acknowledgement is any 2xx response. Receivers must deduplicate by transcription ID. The gateway signs the exact unmodified JSON request-body bytes using HMAC-SHA256.",
                "operationId": "transcription_callback__request_body__callback_url__post",
                "parameters": [
                  {
                    "description": "Unix timestamp in seconds authenticated by the callback signature.",
                    "in": "header",
                    "name": "x-socialguard-timestamp",
                    "required": true,
                    "schema": {
                      "description": "Unix timestamp in seconds authenticated by the callback signature.",
                      "title": "X-Socialguard-Timestamp",
                      "type": "string"
                    }
                  },
                  {
                    "description": "Identifier of the HMAC signing key used for this delivery.",
                    "in": "header",
                    "name": "x-socialguard-webhook-key-id",
                    "required": true,
                    "schema": {
                      "description": "Identifier of the HMAC signing key used for this delivery.",
                      "title": "X-Socialguard-Webhook-Key-Id",
                      "type": "string"
                    }
                  },
                  {
                    "description": "HMAC-SHA256 signature formatted as `sha256=<lowercase-hex>` over ASCII(timestamp) + `.` + the exact JSON request-body bytes.",
                    "in": "header",
                    "name": "x-socialguard-signature",
                    "required": true,
                    "schema": {
                      "description": "HMAC-SHA256 signature formatted as `sha256=<lowercase-hex>` over ASCII(timestamp) + `.` + the exact JSON request-body bytes.",
                      "title": "X-Socialguard-Signature",
                      "type": "string"
                    }
                  }
                ],
                "requestBody": {
                  "content": {
                    "application/json": {
                      "schema": {
                        "discriminator": {
                          "mapping": {
                            "transcription.completed": "#/components/schemas/TranscriptionCompleted",
                            "transcription.failed": "#/components/schemas/TranscriptionFailed"
                          },
                          "propertyName": "type"
                        },
                        "oneOf": [
                          {
                            "$ref": "#/components/schemas/TranscriptionCompleted"
                          },
                          {
                            "$ref": "#/components/schemas/TranscriptionFailed"
                          }
                        ],
                        "title": "Event"
                      }
                    }
                  },
                  "required": true
                },
                "responses": {
                  "200": {
                    "content": {
                      "application/json": {
                        "schema": {}
                      }
                    },
                    "description": "Successful Response"
                  },
                  "422": {
                    "content": {
                      "application/json": {
                        "schema": {
                          "$ref": "#/components/schemas/HTTPValidationError"
                        }
                      }
                    },
                    "description": "Validation Error"
                  }
                },
                "summary": "Transcription callback"
              }
            }
          }
        },
        "description": "Map an accepted HTTP request to the asynchronous submission seam.",
        "operationId": "queueTranscription",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuedTranscription"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authenticated caller is not permitted to perform the operation."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An idempotency key was reused with a different request payload."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The gateway cannot accept more work at this time."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected gateway failure; implementation details are omitted."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The selected model or submission dependency is unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Queue a transcription"
      }
    }
  }
}